TASK: Separate the ``nodeIdentifiers`` utility from ``nodePaths`` and extract common concerns to ``nodeContexts`
